% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{arrange_plots}
\alias{arrange_plots}
\title{arrange_plots}
\usage{
arrange_plots(chart_list, labels = NULL, align_dir = NULL, ...)
}
\arguments{
\item{chart_list}{A list of charts}
}
\description{
Helper function to arrange plots for displaying
}
