% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcn.R
\name{base_chart_specs}
\alias{base_chart_specs}
\title{chartSpecs}
\usage{
base_chart_specs(chart_type = NULL)
}
\arguments{
\item{chartType}{}
}
\value{
character list of chart specifications
}
\description{
Return a list of all specifications for a chart type.
}
