% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcn.R
\name{base_chart_types}
\alias{base_chart_types}
\title{base_chart_types}
\usage{
base_chart_types()
}
\value{
character list of all chart types
}
\description{
Return a list of all chart types that are currently supported in gevitR
}
