% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combo_plots.R
\name{check_link}
\alias{check_link}
\title{helper function to check link between data items}
\usage{
check_link(item_one = NA, item_two = NA)
}
\arguments{
\item{item_two}{}
}
\description{
helper function to check link between data items
}
