% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{combine_charts}
\alias{combine_charts}
\title{Combine multiple charts together}
\usage{
combine_charts(combo_type, charts)
}
\description{
facet_by is for small_multiples
ncol and nrow is for many types general
}
