% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combine_many_types_general}
\alias{combine_many_types_general}
\title{Combine charts with format: many types general}
\usage{
combine_many_types_general(charts, label = NULL)
}
\description{
Combine charts with format: many types general
}
