% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_common_stat_charts.R
\name{common_stats_aesethetics}
\alias{common_stats_aesethetics}
\title{Title}
\usage{
common_stats_aesethetics(gg_chart = NA, title = NA, x_limits = NA,
  y_limits = NA, flip_coord = FALSE, scale_y_cont = NA,
  rm_x_labels = FALSE, rm_y_labels = FALSE, x_labels = FALSE,
  y_labels = FALSE, x_breaks = FALSE, y_breaks = FALSE,
  shrink_plot_margin = FALSE)
}
\arguments{
\item{rm_y_labels}{}
}
\value{
modified gg_chart
}
\description{
Title
}
