% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combo_plots.R
\name{get_axis_info}
\alias{get_axis_info}
\title{Helper function to get axis information from charts}
\usage{
get_axis_info(chart = NULL, align = NULL)
}
\arguments{
\item{align}{}
}
\description{
Helper function to get axis information from charts
}
