% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_genomic.R
\name{get_diff_pos}
\alias{get_diff_pos}
\title{get_diff_post}
\usage{
get_diff_pos(dat)
}
\arguments{
\item{dat}{}
}
\description{
For genomic data, return the positions where all nucleotides are NOT the same
}
