% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{get_limits}
\alias{get_limits}
\title{Get consistent values for the var_name in specified_charts}
\usage{
get_limits(specified_charts, var_name, sum_var = NULL)
}
\arguments{
\item{specified_charts}{A list of chart specifications}

\item{var_name}{A string with the name of the variable to get the limits of. Must be the same name for all data.}

\item{sum_var}{A string with the name of the variable to group for summing}
}
\description{
Get consistent values for the var_name in specified_charts
}
