% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combo_plots.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Helper function to get data out of specific gevitR data type}
\usage{
get_raw_data(item = NA)
}
\arguments{
\item{item}{}
}
\description{
Helper function to get data out of specific gevitR data type
}
