% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{ggplot_scale_info}
\alias{ggplot_scale_info}
\title{ggplot_scale_info}
\usage{
ggplot_scale_info(chart = NULL, spec_list = NULL)
}
\arguments{
\item{chart}{}
}
\description{
Helper function to extract x and y scales from ggplot entity
}
