% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_fcn.R
\name{input_image}
\alias{input_image}
\title{Input Image
A helper function for input_data}
\usage{
input_image(file = NA, desc = NA, asObj = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Input Image
A helper function for input_data
}
