% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{layout_plots}
\alias{layout_plots}
\title{Layout plots}
\usage{
layout_plots(chart_list, labels = NULL, shared_legend = FALSE)
}
\arguments{
\item{chart_list}{A list of charts}
}
\description{
Layout plots
}
