% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcn.R
\name{plot.gevitSpec}
\alias{plot.gevitSpec}
\title{plot.gevitR}
\usage{
\method{plot}{gevitSpec}(specs = NULL, do_not_display = FALSE)
}
\arguments{
\item{specs}{}
}
\value{
a plot that is displayed to the screen
}
\description{
Plot function for gevitR. To use this function, the user only actually has to type plot
}
