% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_chart_types_fcn.R
\name{plot_chart}
\alias{plot_chart}
\title{Plot a basic chart type}
\usage{
plot_chart(data, chart_type, x, y, fill, group, value)
}
\description{
See alternative documentation for input and return values
See alternative documentation for examples
}
