% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_statistical_fcns.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(data, x, y, facet_by)
}
\arguments{
\item{data}{A data frame}

\item{x}{optional categorical variable for x-axis if quantitative value attribute has not been calculated in data}

\item{y}{optional categorical variable for y-axis. If given (along with x_col), will calculate the number of occurences with x_var.}
}
\description{
Plot a heatmap
}
