% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{plot_many_types_general}
\alias{plot_many_types_general}
\title{Many types general plot}
\usage{
plot_many_types_general(...)
}
\arguments{
\item{...}{Any number of lists of arguments to generate a plot}
}
\description{
Many types general plot
}
