% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{plot_simple}
\alias{plot_simple}
\title{Plot a simple chart type}
\usage{
plot_simple(chart_type, data, x_var, y_var, facet_by, fill, group, value)
}
\description{
This function will create a single chart object that can be passed into render_charts() to render.
}
\details{
See alternative documentation for input and return values
See alternative documentation for examples
}
