% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{plot_small_multiples}
\alias{plot_small_multiples}
\title{Small multiples}
\usage{
plot_small_multiples(chart_type, data, facet_by, x = NA, y = NA, z = NA,
  fill = NA, group = NA)
}
\arguments{
\item{chart_type}{A string indicating type of chart to generate. Options are:}

\item{data}{A data frame}

\item{facet_by}{}

\item{x}{optional for: ... required for: ...}

\item{y}{optional for: ... required for: ...}

\item{fill}{optional: ...}

\item{group}{optional: ... not applciable for : ...}
}
\description{
Small multiples
}
