% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{render_charts}
\alias{render_charts}
\title{Render charts --> We have called this combine_charts but maybe render_charts is a better name?}
\usage{
render_charts(charts_list)
}
\description{
Render charts --> We have called this combine_charts but maybe render_charts is a better name?
}
