% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{render_simple}
\alias{render_simple}
\title{Plot a base chart type}
\usage{
render_simple(chart_specs)
}
\description{
This function will create a single chart object that can be passed into render_charts() to render.
}
\details{
See alternative documentation for input and return values
See alternative documentation for examples

TODO: currently, the user must name each of the args that are optional if they don't have them in the right order
}
