% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_combo_plots.R
\name{return_compatible_chart_link}
\alias{return_compatible_chart_link}
\title{Return Compatible Chart Links}
\usage{
return_compatible_chart_link(chart_info = NA, combo_type = NA)
}
\arguments{
\item{chart_info}{}
}
\value{
data frame for compatible charts and their linking variable
}
\description{
Return Compatible Chart Links
}
