% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcn.R
\name{specify_combination}
\alias{specify_combination}
\title{Specify combination}
\usage{
specify_combination(combo_type = NA, base_charts = NA, facet_by = NA,
  link_by = NA, alignment = NA)
}
\arguments{
\item{alignment}{}
}
\description{
Specify combination
}
