% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcns.R
\name{specify_combo}
\alias{specify_combo}
\title{Specify a combination}
\usage{
specify_combo(combo_type, ..., facet_by = NA, link_var = NA,
  link_mark_type = "default", alignment = NA, common_var = NA,
  order = NA)
}
\description{
#'
}
