% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcns.R
\name{specify_reencoding}
\alias{specify_reencoding}
\title{Specify a reencoded mark}
\usage{
specify_reencoding(base_specification, reencode_var,
  mark_type = "default", channel = "colour", override = FALSE)
}
\arguments{
\item{mark_type}{A string that specifies the type of mark to reencode. default depends on the type of chart. Possible strings include: 'default', 'area', 'line', 'point' and 'text'}

\item{channel_type}{A string that specifies the type of channel to reencode. Default is 'colour'. Possible strings include: 'colour' , size', 'shape', 'texture', 'font_face'}
}
\description{
Specify a reencoded mark
}
