% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fcns.R
\name{plot_simple}
\alias{plot_simple}
\title{Plot a base chart type}
\usage{
plot_simple(chart_type, data, x = NA, y = NA, z = NA, stack_by = NA,
  fill = NA, group = NA, title = NA, path, category, cluster_vars = NA,
  tip_var = NA, comparisons, layout = "default", proportional = FALSE,
  reference_vector, reference_var, key, value, date, start = NA, end = NA,
  names = NA, events = NA, rownames = NA, lat_var = NA, long_var = NA,
  flip_coord = FALSE, rm_y_labels = FALSE, rm_x_labels = FALSE,
  colour_var = NA, colour_scale = NA, link_mark_type = NA,
  x_limits = NA, y_limits = NA)
}
\description{
This function will create a single chart object that can be passed into layout_charts() to layout.
}
\details{
See alternative documentation for input and return values
See alternative documentation for examples

TODO: currently, the user must name each of the args that are optional if they don't have them in the right order
}
