% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcns.R
\name{specify_base}
\alias{specify_base}
\title{Specify requirements to make a base chart}
\usage{
specify_base(chart_type, data, x, y, z, ...)
}
\arguments{
\item{chart_type}{}
}
\description{
Specify requirements to make a base chart
}
