% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_fcn.R
\name{specify_base}
\alias{specify_base}
\title{specify_base}
\usage{
specify_base(chart_type = NULL, data = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{
a specification for a base chart type that will be read by gevitR::plot
}
\description{
Specify a base chart type
}
